+++ CHANGELOG.md
@@ -2,3 +2,8 @@
 
 - Initial version.
 - @Singleton added
+  
+## 0.0.1-dev.2
+ - improved code
+ - abstract and sealed class detection
+ - better error reporting